chore : update some dto

+++ server/user_service/src/main/java/com/final_project_clinic/user/dto/PatientSaveDTO.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class PatientSaveDTO {
     private UUID user_id;
-    private Integer nik;
+    private Long nik;
     private String phoneNumber;
     private String address;
     private String gender;
